requirements.txt
odoo-test-helper
python-barcode
barcodenumber
